feat: add flag to disable offsettoid in Offsettoid class

+++ docarray/array/storage/annlite/getsetdel.py
@@ -46,15 +46,25 @@ class GetSetDelMixin(BaseGetSetDelMixin):
         self._annlite.close()
 
     def _load_offset2ids(self):
-        self._offsetmapping = OffsetMapping(
-            data_path=self._config.data_path, in_memory=False
-        )
-        self._offsetmapping.create_table()
-        self._offset2ids = Offset2ID(self._offsetmapping.get_all_ids())
+        if self._list_like:
+            self._offsetmapping = OffsetMapping(
+                data_path=self._config.data_path, in_memory=False
+            )
+            self._offsetmapping.create_table()
+            self._offset2ids = Offset2ID(
+                self._offsetmapping.get_all_ids(),
+                list_like=self._list_like,
+            )
+        else:
+            self._offset2ids = Offset2ID(
+                [],
+                list_like=self._list_like,
+            )
 
     def _save_offset2ids(self):
-        self._offsetmapping.drop()
-        self._offsetmapping.create_table()
-        self._offsetmapping._insert(
-            [(i, doc_id) for i, doc_id in enumerate(self._offset2ids.ids)]
-        )
+        if self._list_like:
+            self._offsetmapping.drop()
+            self._offsetmapping.create_table()
+            self._offsetmapping._insert(
+                [(i, doc_id) for i, doc_id in enumerate(self._offset2ids.ids)]
+            )
